Install npm, add fix after review
