+++ .eslintrc.yml
@@ -8,5 +8,8 @@ env:
   es2017: true
   browser: true
 
-extends:
-  "htmlacademy/vanilla"
+extends: "htmlacademy/vanilla"
+
+rules:
+  semi: error
+
