+++ js/main.js
@@ -1,7 +1,6 @@
 import { generateRandomAds } from './data.js';
-import { addFormListeners } from './form.js';
-import { addPins, mapInit } from './map.js';
-import { adFormActivationToggle } from './forms-activation.js';
+import { addFormListeners, adFormsActivationToggle } from './form.js';
+import { addPins, mapInit, addMainPinMarker } from './map.js';
 import { generateAdElement } from './similar-ads.js';
 
 const ads = generateRandomAds();
@@ -9,9 +8,10 @@ const pinsData = ads.map((ad) => {
   return {
     popupContent: generateAdElement(ad),
     location: ad.location,
-  }
+  };
 });
 
-mapInit(adFormActivationToggle);
+mapInit(adFormsActivationToggle);
 addFormListeners();
 addPins(pinsData);
+addMainPinMarker();
